Рефакторинг reducer
